+++ il_bank_validator/__init__.py
@@ -1,4 +1,4 @@
-from .validator import validate_israeli_bank_account
 from .exceptions import InvalidBankAccountException
+from .validator import validate_israeli_bank_account
 
 __all__ = ["validate_israeli_bank_account", "InvalidBankAccountException"]