+++ il_bank_validator/exceptions.py
@@ -1,3 +1,4 @@
 class InvalidBankAccountException(Exception):
     """Raised when a bank account input is invalid."""
+
     pass